BUTTONS: Rework and expand button handling
This adds functionality for addition button options.  Adds short and
long hold options for the A, B, and trigger buttons, plus long hold
options for the gimbal preset buttons.  This also significantly cleans
up the handling, making it more robust.  Junk and unused functions are
dumped.  This is required to marry up with similar updates to the Artoo
STM32.  As such, you do need to have Open Solo on both the controller
and solo.  Also adds additional logging.  Additional commits to follow
make corresponding button handling changes to the smart shots, which are
also required.